refactor(salariesInput): break down SalariesInput into smaller components
