+++ studio/components/salariesInput/salariesInput.module.css
